Refactor Running Pace API and add world records endpoint
This commit refactors the Running Pace API by moving the logic from the pace_calculator module to the new core module. The main changes include:

    - Renaming the pace_calculator module to calculator and moving it to the core subpackage.
    - Moving the OFFICIAL_DISTANCES constant to the models module.
    - Creating a new scrapper module in the core subpackage to handle scraping of world records data.
    - Adding a new endpoint /get_world_records to retrieve world records data from the World Athletics website.
    - Updating the imports in the test modules to reflect the new module structure.
    - Updating the entry point in run.sh to use the new main module.

This refactor improves the modularity and organization of the code, making it easier to maintain and extend in the future. The addition of the world records endpoint provides a new feature for users to access up-to-date world records data.